% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/craft.R
\name{craft}
\alias{craft}
\title{What craft do you do?}
\usage{
craft(art, supplies)
}
\arguments{
\item{art}{A character of your artistic hobby!}

\item{supplies}{A character of the supplies you use}
}
\value{

}
\description{
Answer the question what craft you do!
}
