% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getExamplesAPI}
\alias{getExamplesAPI}
\title{Word usage examples, bound to the given word id in synsets}
\usage{
getExamplesAPI(word_id = "")
}
\arguments{
\item{id}{The input word id}
}
\value{
A list of objects holding examples for a given word id
}
\description{
Word usage examples, bound to the given word id in synsets
}
\examples{
word_id <- getIdByWordAPI("машина")
examples2df(getExamplesAPI(word_id))
}
\seealso{
\code{\link{getRawExamplesAPI}}
}

