% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIdByWord}
\alias{getIdByWord}
\title{Get the ID for a given word}
\usage{
getIdByWord(word = "")
}
\arguments{
\item{word}{The input word}
}
\value{
ID
}
\description{
Get the ID for a given word
}
\examples{
getIdByWord('музыка')
}
\seealso{
\code{\link{getWordById}, \link{getIdByWordAPI}}
}

