% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getIdByWordAPI}
\alias{getIdByWordAPI}
\title{Get the ID for a given word}
\usage{
getIdByWordAPI(word = "")
}
\arguments{
\item{word}{The input word}
}
\value{
ID
}
\description{
Get the ID for a given word
}
\examples{
\dontrun{
getIdByWordAPI('house')
}
}
\references{
Dmitriy Ustalov. YARN API. \url{http://nlpub.ru/YARN/API}
}
\seealso{
\code{\link{getIdByWord}}
}

