% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRawExamplesAPI}
\alias{getRawExamplesAPI}
\title{Примеры употребления из «сырых» словарей для слова с указанным идентификатором id}
\usage{
getRawExamplesAPI(word_id = "")
}
\arguments{
\item{id}{Id}
}
\description{
Примеры употребления из «сырых» словарей для слова с указанным идентификатором id
}

