% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSynonymsAPI}
\alias{getSynonymsAPI}
\title{Get synonymous for a given word id}
\usage{
getSynonymsAPI(word_id)
}
\arguments{
\item{word_id}{Id}
}
\value{
The output is an array with objects of class Word
}
\description{
Get synonymous for a given word id
}
\examples{
word_id <- getIdByWordAPI("машина")
synonymous <- words2df(getSynonymsAPI(word_id))
}

