% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSynsets}
\alias{getSynsets}
\title{Get synsets for a given word}
\usage{
getSynsets(word)
}
\arguments{
\item{word}{The input word}
}
\value{
A list of synsets
}
\description{
Get synsets for a given word
}
\examples{
\dontrun{
getSynsets('house')
}
}
\seealso{
\code{\link{getSynsetsAPI}}
}

