% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSynsetsAPI}
\alias{getSynsetsAPI}
\title{Get synsets for a given word id (require registration)}
\usage{
getSynsetsAPI(page = 1, id = "")

getSynsetsAPI(page = 1, id = "")
}
\arguments{
\item{page}{Page}

\item{id}{Id}

\item{word_id}{}
}
\description{
Get synsets for a given word id (require registration)

Get synsets
}

