% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSynsetsByIdAPI}
\alias{getSynsetsByIdAPI}
\title{Get a full list of known synsets with page breakdown}
\usage{
getSynsetsByIdAPI(page = 1, id = "")
}
\arguments{
\item{page}{Page number}

\item{id}{Synset id}
}
\value{
A list of synsets
}
\description{
Get a full list of known synsets with page breakdown
}
\examples{
getSynsetsByIdAPI(page=1)
getSynsetsByIdAPI(id=1)
}
\references{
Dmitriy Ustalov. YARN API. \url{http://nlpub.ru/YARN/API}
}

