% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWordById}
\alias{getWordById}
\title{Get the word for a given id}
\usage{
getWordById(word_id)
}
\arguments{
\item{word_id}{The input word id}
}
\value{
The word
}
\description{
Get the word for a given id
}
\examples{
getWordById(131)
}
\seealso{
\code{\link{getIdByWord}}
}

