% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWordDefinitionsAPI}
\alias{getWordDefinitionsAPI}
\title{Definitions, bound to the given word id in synset}
\usage{
getWordDefinitionsAPI(word_id = "")
}
\arguments{
\item{id}{The input word id}
}
\value{
A list of objects holding definitions for a given word id
}
\description{
Definitions, bound to the given word id in synset
}
\examples{
word_id <- getIdByWordAPI("машина")
definitions2df(getWordDefinitionsAPI(word_id))
}
\seealso{
\code{\link{getWordRawDefinitionsAPI}}
}

