% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWordRawDefinitionsAPI}
\alias{getWordRawDefinitionsAPI}
\title{Definitions from the "raw" dictionaries for a given word id}
\usage{
getWordRawDefinitionsAPI(word_id = "")
}
\arguments{
\item{id}{The input word id}
}
\value{
A list of objects holding definitions for a given word id
}
\description{
Definitions from the "raw" dictionaries for a given word id
}
\examples{
word_id <- getIdByWordAPI("машина")
definitions2df(getWordRawDefinitionsAPI(word_id))
}
\seealso{
\code{\link{getWordDefinitionsAPI}}
}

