% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWords}
\alias{getWords}
\title{Get the words in a synset}
\usage{
getWords(synset)
}
\arguments{
\item{synset}{The synset whose words are returned}
}
\value{
A character vector holding the words
}
\description{
Get the words in a synset
}
\examples{
\dontrun{
synsets <- getSynsets('house')
getWords(synsets[[1]])
}
}

