% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWordsAPI}
\alias{getWordsAPI}
\title{Получить список всех слов}
\usage{
getWordsAPI(page = 1, start = "")
}
\arguments{
\item{page}{Номер страницы}

\item{start}{Подстрока}
}
\value{
На выходе получается массив с объектами класса Word
}
\description{
Поиск по известным словам с постраничной разбивкой. Слово начинается с подстроки word.
}
\examples{
word_objs <- getWordsAPI(start='машина')
words <- words2df(words_obj)
}

