% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCMDS.R
\name{plotCMDS}
\alias{plotCMDS}
\title{Plot classical MDS of dataset}
\usage{
plotCMDS(obj, comp = 1:2, distFun = dist, distMethod = "euclidian",
  n = NULL, samples = TRUE, ...)
}
\arguments{
\item{obj}{ExpressionSet object or objrix.}

\item{comp}{Which components to display.}

\item{distFun}{Distance function, default is dist.}

\item{distMethod}{The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}

\item{n}{Number of features to make use of in calculating your distances.}

\item{samples}{Perform on samples or genes.}

\item{...}{Additional plot arguments.}
}
\value{
coordinates
}
\description{
This function plots the MDS coordinates for the "n" features of interest. Potentially uncovering batch
effects or feature relationships.
}
\examples{
\donttest{data(skin)
res = plotCMDS(skin)
library(calibrate)
textxy(X=res[,1],Y=res[,2],labs=rownames(res))}
}

