% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweater.R
\name{sweater}
\alias{sweater}
\title{Tell Me About Your Sweater!}
\usage{
sweater(design, fiber)
}
\arguments{
\item{design}{This character is the name of the sweater design}

\item{fiber}{This character is the fiber you used to knit your sweater (ie alpaca)}
}
\value{

}
\description{
Tell Me About Your Sweater!
}
