% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wacky.R
\name{theme_wacky}
\alias{theme_wacky}
\title{Wacky Theme for Scatterplots}
\usage{
theme_wacky()
}
\value{

}
\description{
This theme adds some wacky effects to your data visualization. If you want your eyeballs to bleed, choose this theme!
}
\examples{

}
