% Generated by roxygen2 (4.0.2): do not edit by hand
\name{words2df}
\alias{words2df}
\title{Convert words to a Data Frame}
\usage{
words2df(word_objects)
}
\arguments{
\item{word_objects}{A list with objects of class Word}
}
\value{
Returns a data frame
}
\description{
Convert words to a Data Frame
}
\seealso{
\code{\link{getWordsAPI}}
}

