\name{yarn-package}
\alias{yarn-package}
\alias{yarn}
\docType{package}
\title{
YARN interface
}
\description{
An interface to YARN API. YARN (Yet Another RussNet) is an electronic thesaurus of Russian language, open resource for free access. Words are ordered by semantic proximity.
}
\details{
\tabular{ll}{
Package: \tab yarn\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-05\cr
License: \tab GPL-3\cr
}
}
\author{
Dmitriy Sorokin

Maintainer: Dmitriy Sorokin <dementiy@yandex.ru>
}
\references{
Braslavski P., Ustalov D., Mukhin M. A Spinning Wheel for YARN: User Interface for a Crowdsourced Thesaurus // Proceedings of the Demonstrations at the 14th Conference of the European Chapter of the Association for Computational Linguistics. — Gothenburg, Sweden : Association for Computational Linguistics, 2014. — P. 101–104
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ wordnet, thesaurus  }