% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarn.R
\name{yarn_add}
\alias{yarn_add}
\alias{yarn_install}
\title{Yarn Install}
\usage{
yarn_add(
  ...,
  scope = c("dev", "prod", "peer", "tilde", "exact", "global", "optional")
)

yarn_install()
}
\arguments{
\item{...}{Names of packages to install.}

\item{scope}{Scope of the installation of the packages.}
}
\description{
Yarn Install
}
\section{Functions}{
\itemize{
\item \code{yarn_install}: Install dependencies from project.
}}

