% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDefinitionsAPI}
\alias{getDefinitionsAPI}
\title{Get a list of all definitions with page breakdown}
\usage{
getDefinitionsAPI(page = 1, id = "")
}
\arguments{
\item{page}{The page number}

\item{id}{Id}
}
\value{
The output is an array with objects of class Definition
}
\description{
Get a list of all definitions with page breakdown
}
\examples{
definition_objs <- getDefinitionsAPI(page=1)
definitions <- definitions2df(definition_objs)

definition_obj <- getDefinitionsAPI(id=1)
definition <- definitions2df(definition_obj)
}

