% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getExamplesAPI}
\alias{getExamplesAPI}
\title{Примеры употребления слова, привязанные в синсетах к слову с указанным идентификатором id}
\usage{
getExamplesAPI(word_id = "")
}
\arguments{
\item{id}{Id}
}
\description{
Примеры употребления слова, привязанные в синсетах к слову с указанным идентификатором id
}

