% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRawExamplesAPI}
\alias{getRawExamplesAPI}
\title{Word usage examples from the "raw" dictionaries for a given word id}
\usage{
getRawExamplesAPI(word_id = "")
}
\arguments{
\item{id}{The input word id}
}
\value{
A list of objects holding examples for a given word id
}
\description{
Word usage examples from the "raw" dictionaries for a given word id
}
\examples{
word_id <- getIdByWordAPI("машина")
examples2df(getRawExamplesAPI(word_id))
}
\seealso{
\code{\link{getExamplesAPI}}
}

