% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRawSynonymsAPI}
\alias{getRawSynonymsAPI}
\title{Get raw synonymous for a given word id}
\usage{
getRawSynonymsAPI(word_id)
}
\arguments{
\item{word_id}{Id}
}
\value{
A list holding the synonums for a given word id
}
\description{
Get raw synonymous for a given word id
}
\examples{
\dontrun{
word_id <- getIdByWordAPI("house")
words2df(getRawSynonymsAPI(word_id))
}
}
\seealso{
\code{\link{getSynonymsAPI}}
}

