% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRawSynonymsAPI}
\alias{getRawSynonymsAPI}
\title{Get raw synonymous for a given word id (require registration)}
\usage{
getRawSynonymsAPI(word_id)
}
\arguments{
\item{word_id}{Id}
}
\description{
Get raw synonymous for a given word id (require registration)
}

