% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSynonymsAPI}
\alias{getSynonymsAPI}
\title{Get synonymous for a given word id}
\usage{
getSynonymsAPI(word_id)
}
\arguments{
\item{word_id}{Id}
}
\value{
A list holding the synonums for a given word id
}
\description{
Get synonymous for a given word id
}
\examples{
word_id <- getIdByWordAPI("машина")
words2df(getSynonymsAPI(word_id))
}
\references{
Dmitriy Ustalov. YARN API. \url{http://nlpub.ru/YARN/API}
}
\seealso{
\code{\link{synonyms}, \link{getRawSynonymsAPI}}
}

