% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSynonymsAPI}
\alias{getSynonymsAPI}
\title{Get synonymous for a given word id (require registration)}
\usage{
getSynonymsAPI(word_id)
}
\arguments{
\item{word_id}{Id}
}
\description{
Get synonymous for a given word id (require registration)
}

