% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSynsetsAPI}
\alias{getSynsetsAPI}
\title{Get synsets for a given word id}
\usage{
getSynsetsAPI(word_id)
}
\arguments{
\item{word_id}{Id}
}
\value{
The output is an array with objects of class Synset
}
\description{
Get synsets for a given word id
}

