% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getSynsetsAPI}
\alias{getSynsetsAPI}
\title{Get synsets for a given word id}
\usage{
getSynsetsAPI(word_id)
}
\arguments{
\item{word_id}{The input word id}
}
\value{
A list of synsets
}
\description{
Get synsets for a given word id
}
\examples{
\dontrun{
getSynsets(131)
}
}
\references{
Dmitriy Ustalov. YARN API. \url{http://nlpub.ru/YARN/API}
}

