% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWordDefinitionsAPI}
\alias{getWordDefinitionsAPI}
\title{Определения, привязанные в синсетах к слову с указанным идентификатором id}
\usage{
getWordDefinitionsAPI(word_id = "")
}
\arguments{
\item{id}{Id}
}
\description{
Определения, привязанные в синсетах к слову с указанным идентификатором id
}

