% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWordRawDefinitionsAPI}
\alias{getWordRawDefinitionsAPI}
\title{Определения из «сырых» словарей для слова с указанным идентификатором id}
\usage{
getWordRawDefinitionsAPI(word_id = "")
}
\arguments{
\item{id}{Id
word_id <- getIdByWordAPI("машина")
definitions <- definitions2df(getWordRawDefinitionsAPI(word_id))}
}
\description{
Определения из «сырых» словарей для слова с указанным идентификатором id
}

