% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWordsAPI}
\alias{getWordsAPI}
\title{Get a list of all words from a YARN dictionary}
\usage{
getWordsAPI(page = 1, q = "")
}
\arguments{
\item{page}{Page number}

\item{q}{The word starts with the substring `word`}
}
\value{
A list with objects of class Word
}
\description{
Get a list of all words from a YARN dictionary
}
\examples{
\dontrun{
word_objs <- getWordsAPI(q = 'house')
words2df(word_objs)

word_objs <- getWordsAPI(page = 2)
words2df(word_objs)
}
}
\references{
Dmitriy Ustalov. YARN API. \url{http://nlpub.ru/YARN/API}
}

