% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getWordsAPI}
\alias{getWordsAPI}
\title{A list of all words}
\usage{
getWordsAPI(page = 1, q = "")
}
\arguments{
\item{page}{Page number}

\item{q}{The word starts with the substring `word`}
}
\value{
The output is an array with objects of class Word
}
\description{
The full list of known words with page breakdown.
}
\examples{
word_objs <- getWordsAPI(q='машина')
words <- words2df(words_obj)
}

