% Generated by roxygen2 (4.0.2): do not edit by hand
\name{synonyms}
\alias{synonyms}
\title{Поиск синонимов для указанного слова}
\usage{
synonyms(word)
}
\arguments{
\item{word}{Слово, для которого производится поиск синонимов}
}
\value{
Список найденных синонимов
}
\description{
Краткое описание работы функции
}
\examples{
car_synonyms <- synonyms("машина")
}
\seealso{
\code{\link{synonymsAPI}}
}

