% Generated by roxygen2 (4.0.2): do not edit by hand
\name{synonyms}
\alias{synonyms}
\title{Get synonymous for a given word id}
\usage{
synonyms(word)
}
\arguments{
\item{word}{The input word}
}
\value{
A character vector holding the synonyms for the given word id
}
\description{
Get synonymous for a given word id
}
\examples{
car_synonyms <- synonyms("машина")
}
\seealso{
\code{\link{getSynonymsAPI}}
}

