% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{yarn-package}
\alias{yarn}
\alias{yarn-package}
\title{YARN Interface}
\description{
An interface to YARN API. YARN (Yet Another RussNet) is an electronic thesaurus of Russian language, open resource for free access. Words are ordered by semantic proximity.
}
\author{
Dmitriy Sorokin
}
\references{
Braslavski P., Ustalov D., Mukhin M. A Spinning Wheel for YARN: User Interface for a Crowdsourced Thesaurus
// Proceedings of the Demonstrations at the 14th Conference of the European Chapter of the Association for Computational Linguistics.
 - Gothenburg, Sweden : Association for Computational Linguistics, 2014. - P. 101-104

Yet Another RussNet (2013). \url{http://russianword.net/}
}
\keyword{thesarus}
\keyword{wordnet}

