% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarn.R
\name{yarn}
\alias{yarn}
\title{Knitting for Friends}
\usage{
yarn(friend_name, yarn_yardage, garment)
}
\arguments{
\item{friend_name}{This is a character argument that is your friend's name}

\item{yarn_yardage}{This is a number that tells you yarn yardage}

\item{garment}{A character containing a knit garment}
}
\value{

}
\description{
This is a function that tells a friend what you would knit them using the yarn yardage you have!
}
